requirements.txt
GroupyAPI==0.10.3
atomicwrites==1.3.0
google-api-python-client==1.7.11